feat(#642): WithoutLines.java
