+++ src/main/java/org/eolang/jeo/matchers/WithoutLines.java
@@ -26,50 +26,42 @@ package org.eolang.jeo.matchers;
 import com.jcabi.xml.XML;
 import com.jcabi.xml.XMLDocument;
 import org.cactoos.Scalar;
-import org.hamcrest.Description;
-import org.hamcrest.TypeSafeMatcher;
 import org.xembly.Directives;
 import org.xembly.ImpossibleModificationException;
 import org.xembly.Xembler;
 
 /**
- * Matcher to check if the received XML document is the same as the expected one.
- * Smart comparison of XML documents that ignores 'line' attributes.
+ * The same XML, but without lines.
+ *
  * @since 0.6
  */
-public final class SameXml extends TypeSafeMatcher<String> {
+public final class WithoutLines implements Scalar<XML> {
 
     /**
-     * Expected XML document.
+     * The original.
      */
-    private final String expected;
-
-    /**
-     * Constructor.
-     * @param xml Expected XML document.
-     */
-    public SameXml(final XML xml) {
-        this(xml.toString());
-    }
+    private final XML original;
 
     /**
      * Constructor.
-     * @param expected Expected XML document.
+     * @param orgnl Original XML
      */
-    public SameXml(final String expected) {
-        this.expected = expected;
+    public WithoutLines(final XML orgnl) {
+        this.original = orgnl;
     }
 
     @Override
-    public boolean matchesSafely(final String item) {
-        return new WithoutLines(new XMLDocument(this.expected)).value()
-            .equals(new WithoutLines(new XMLDocument(item)).value());
-    }
-
-    @Override
-    public void describeTo(final Description description) {
-        description.appendText("XML documents is not the same.")
-            .appendText("Expected:\n")
-            .appendText(this.expected);
+    public XML value() {
+        try {
+            return new XMLDocument(
+                new Xembler(
+                    new Directives().xpath(".//o[@line]/@line").remove()
+                ).apply(this.original.node())
+            );
+        } catch (final ImpossibleModificationException exception) {
+            throw new IllegalStateException(
+                "Failed to remove 'line' attributes", exception
+            );
+        }
     }
 }
